fix : rate limit service to handle the new flow
